app/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
APScheduler==3.9.1.post1
attrs==22.1.0
billiard==3.6.4.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crashtest==0.3.1
distlib==0.3.6
dnspython==2.2.1
dulwich==0.20.50
filelock==3.8.0
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-RESTful==0.3.9
greenlet==2.0.1
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.0
keyring==23.11.0
kombu==5.2.4
lockfile==0.12.2
MarkupSafe==2.1.1
more-itertools==9.0.0
msgpack==1.0.4
mypy==0.991
mypy-extensions==0.4.3
packaging==21.3
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.4
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
xattr==0.9.9
zipp==3.10.0
